+++ aio_pika/adapter.py
@@ -129,6 +129,7 @@ class AsyncioConnection(base_connection.BaseConnection):
         self.loop = loop or asyncio.get_event_loop()
         self.ioloop = IOLoopAdapter(self.loop)
         self.channel_cleanup_callback = None
+        self.channel_cancel_callback = None
 
         super().__init__(parameters, on_open_callback,
                          on_open_error_callback,
@@ -182,9 +183,15 @@ class AsyncioConnection(base_connection.BaseConnection):
         finally:
             super()._on_channel_cleanup(channel)
 
+    def _on_channel_cancel(self, channel):
+        if self.channel_cancel_callback:
+            self.channel_cancel_callback(channel)
+
     def _create_channel(self, channel_number, on_open_callback):
         log.debug('Creating channel %s', channel_number)
-        return Channel(self, channel_number, on_open_callback)
+        channel = Channel(self, channel_number, on_open_callback)
+        channel.add_on_cancel_callback(lambda method_frame: self._on_channel_cancel(channel))
+        return channel
 
 
 class Channel(channel.Channel):