+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.5.0
+-----
+
+* Ability to reconnect on Channel.Close
+* Ability to reconnect on Channel.Cancel
+
 2.4.0
 -----